remove linters errors
