--- Student.rb
@@ -1,11 +1,0 @@
-module Student
-	module Student
-		def initialize(classroom)
-			@classroom = classroom
-		end
-	
-		def play_hooky
-			"¯\(ツ)/¯"
-		end  
-	end
-end
