+++ student.rb
@@ -1,0 +1,9 @@
+module Student
+  def initialize(classroom)
+    @classroom = classroom
+  end
+
+  def play_hooky
+    "¯\(ツ)/¯"
+  end
+end
