+++ teacher.rb
@@ -1,9 +1,7 @@
 module Teacher
-  module Teacher
     def initialize(specialization)
-        @specialization=specialization
+      @specialization = specialization
       super
     end
+  end
 
-  end 
-end
